Updated GUI
Implemented the direction indicator for the virtual representation of the bot. Added autonomous selection buttons to the Auton screen (these do not do anything yet). Added a BRAIN_SCREEN_GAME_MODE macro to drivetrain.hpp. When BRAIN_SCREEN_GAME_MODE is defined, instead of moving the bot the controller moves the virtual bot around the virtual field on the GUI.