Updated GUI, Auton Selection Structure
Replaced DISPLAY_DEBUG_LEVEL with DISPLAY_DEBUG. When defined, the gui always fully shows. When not defined, only the auton selection tab will show once the bot is enabled. For auton selection, instead of using an enum class to keep track of the selected auton and then using a switch statement in the autonomous function, function pointers are used instead.