Updated drivetrain files
Removed Drivetrain::State class and all usages of State, as we will very likely only use linear drives this season. For the move to point command (operator>>), if a heading is provided to the passed Point, the bot will turn to that heading after reaching the desired position on the field.